refactor: rename worker commands to driver-commands
