+++ src/worker/driver-commands/change-price.ts
@@ -1,6 +1,6 @@
 import { createAppRequestResponsePair } from "@app/messaging";
 
-export const changePriceCommand = createAppRequestResponsePair<
+export const changePriceDriverCommand = createAppRequestResponsePair<
   "CHANGE_PRICE",
   { newPrice: string; offerEditUrl: string }
 >("CHANGE_PRICE");
