+++ src/worker/driver-commands/get-offers.ts
@@ -1,8 +1,8 @@
 import type { Offer } from "@app/marketplaces/common/messaging";
 import { createAppRequestResponsePair } from "@app/messaging";
 
-export const getOffersCommand = createAppRequestResponsePair<
-  "GET_OFFERS",
+export const getOffersDriverCommand = createAppRequestResponsePair<
+  "DRIVER/GET_OFFERS",
   void,
   Offer[]
->("GET_OFFERS");
+>("DRIVER/GET_OFFERS");
