+++ src/worker/index.ts
@@ -1,2 +1,2 @@
 export * from "./state";
-export * from "./commands";
+export * from "./driver-commands";
